Generate named nested types with doc comments
Generate named nested types instead of anonymous inline expansions for
TypeScript. This makes it possible to add doc comments to these nested types,
which will be available in the API docs.

The nested types are emitted under a top-level `types` module in `input` or
`output` submodules, similar to the schemed used by the Kubernetes provider,
which is done to avoid drowning out the types that users actually want to use in
the main modules.

So, for example, the type of the `website` property of `aws.s3.BucketArgs` is
`aws.types.input.s3.BucketWebsite`.

There are a couple differences here compared to the Kubernetes provider:

 1. The input/output submodules are truly modules, whereas in the Kubernetes
    provider they are TypeScript namespaces. The difference is outlined at
    https://www.typescriptlang.org/docs/handbook/namespaces-and-modules.html.
    The reason we use modules is because it's easier to import custom overlays
    only where needed using actual modules (it's not possible to nest imports
    within namespaces) and the docs site currently isn't capable of outputting
    types in namespaces (though, this will be fixed soon).

 2. The Kubernetes provider puts `Args` types in `k8s.types.input`, but these
    have already been emitted in the main modules here.

The names are generated based on the resource name and property chain. So the
name of the type used for the `objectLockConfiguration` property on
`aws.s3.BucketArgs` is `aws.types.input.s3.BucketObjectLockConfiguration`, and
its `rule` property is typed as a further nested type named
`aws.types.input.s3.BucketObjectLockConfigurationRule`.

The name of types used in arrays are made singular, so if a property on
`FooArgs` is named `rules` and is an array, its type would be `FooRule[]`.